+++ CHANGELOG.md
@@ -4,6 +4,9 @@ Version 0.5.6
 **Improvements**
 
 - Bugfix: list MacOS application's codesigning certificates as well in `keychain list-certificates`.
+- Bugfix: in `xcode-project use-profiles` search for provision profiles with `.provisionprofile` extension as well.
+- Bugfix: handle provisioning profiles entitlements keys with prefixes e.g. `com.apple.application-identifier`
+- Bugfix: do not default to iphoneos sdk when setting code signing information
 
 Version 0.5.5
 -------------